requirements.txt
Django~=5.2.8
djangorestframework~=3.16.1
psycopg2-binary>=2.9
gunicorn>=21.0